app/requirements.txt
Django
djangorestframework
django-filter
django-cors-headers
djangorestframework-simplejwt
dj-rest-auth
gunicorn
Pillow
django-import-export